requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
gunicorn>=21.2,<22.0
psycopg2-binary>=2.9,<3.0
